# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pets App API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get list of the events.",
                "operationId": "9d193c96b81abc746c99eff78d54737d",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/EventShortResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Create a new event.",
                "operationId": "7d4658700b83072ed94175c8f1bac8a3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EventStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventFullResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/events/{event}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get an event.",
                "operationId": "ff2b02c83ff5f2955a7dfe3f703461bc",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/EventFullResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Delete an event.",
                "operationId": "725a2330cf730965879db5b5bc43b0c3",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824ca"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful response"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Events"
                ],
                "summary": "Update an event.",
                "operationId": "0a2783f6ecd6226fe3859046c3522e39",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824ca"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EventStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventFullResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/notes": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Get list of the notes.",
                "operationId": "9a959e36a067705923b1403aa9ce6bde",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/NoteShortResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Notes"
                ],
                "summary": "Create a new note.",
                "operationId": "6149b28393a9a434f11d4ec7e605cce8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NoteStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoteFullResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/notes/{note}": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Get a note.",
                "operationId": "f5701b3b3b894427bec8546f63e2a0f3",
                "parameters": [
                    {
                        "name": "note",
                        "in": "path",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824ca"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoteFullResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notes"
                ],
                "summary": "Delete the note.",
                "operationId": "f26de33ac1d36467b48d54674c3c2afc",
                "parameters": [
                    {
                        "name": "note",
                        "in": "path",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824ca"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful response"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Notes"
                ],
                "summary": "Update the note.",
                "operationId": "9dfc1c405ebd2ecb37b09de2692d9e2a",
                "parameters": [
                    {
                        "name": "note",
                        "in": "path",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824ca"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NoteStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoteFullResource"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "EventStoreRequest": {
                "properties": {
                    "animal_id": {
                        "type": "string",
                        "example": "995037a6-60b3-4055-aa14-3513aa9824ca"
                    },
                    "category_ids": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "title": {
                        "type": "string",
                        "example": "Some title"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "starts_at": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "ends_at": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "repeat": {
                        "type": "string",
                        "example": "never"
                    },
                    "whole_day": {
                        "type": "boolean",
                        "example": false
                    },
                    "only_this": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "NoteStoreRequest": {
                "properties": {
                    "animal_id": {
                        "type": "string",
                        "example": "995037a6-60b3-4055-aa14-3513aa9824ca"
                    },
                    "category_ids": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "title": {
                        "type": "string",
                        "example": "Some title"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "AnimalFullResource": {
                "properties": {
                    "uuid": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string",
                        "example": "Fluffy"
                    },
                    "birth_date": {
                        "type": "string",
                        "example": "2010-10-02"
                    },
                    "type": {
                        "type": "string",
                        "example": "Type"
                    },
                    "breed": {
                        "type": "string",
                        "example": "Breed"
                    },
                    "sex": {
                        "type": "string",
                        "example": "Male"
                    },
                    "weight": {
                        "type": "string",
                        "example": 5
                    },
                    "avatar": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "thumb": {
                                    "type": "string",
                                    "example": null,
                                    "nullable": true
                                },
                                "full": {
                                    "type": "string",
                                    "example": null,
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    },
                    "activiry": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "notes": {
                                    "type": "integer",
                                    "example": 2
                                },
                                "events": {
                                    "type": "integer",
                                    "example": 12
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "DictionaryTypedResource": {
                "required": [
                    "uuid",
                    "name"
                ],
                "properties": {
                    "uuid": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "EventFullResource": {
                "properties": {
                    "uuid": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string",
                        "example": "Some title"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "starts_at": {
                        "type": "string",
                        "example": "2023-06-10 10:00:00"
                    },
                    "ends_at": {
                        "type": "string",
                        "example": "2023-06-12 18:00:00"
                    },
                    "repeat": {
                        "type": "string",
                        "example": "never"
                    },
                    "whole_day": {
                        "type": "boolean",
                        "example": false
                    },
                    "animal": {
                        "$ref": "#/components/schemas/AnimalFullResource"
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DictionaryTypedResource"
                        }
                    }
                },
                "type": "object"
            },
            "EventShortResource": {
                "properties": {
                    "uuid": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string",
                        "example": "Some title"
                    },
                    "starts_at": {
                        "type": "string",
                        "example": "2023-06-10 10:00:00"
                    },
                    "ends_at": {
                        "type": "string",
                        "example": "2023-06-12 18:00:00"
                    },
                    "repeat": {
                        "type": "string",
                        "example": "never"
                    },
                    "whole_day": {
                        "type": "boolean",
                        "example": false
                    },
                    "animal": {
                        "type": "string"
                    },
                    "categories": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "NoteFullResource": {
                "properties": {
                    "uuid": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "descrpition": {
                        "type": "string",
                        "nullable": true
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DictionaryTypedResource"
                        }
                    },
                    "animal": {
                        "$ref": "#/components/schemas/AnimalFullResource"
                    },
                    "last_action_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "NoteShortResource": {
                "properties": {
                    "uuid": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "descrpition": {
                        "type": "string",
                        "nullable": true
                    },
                    "categories": {
                        "type": "string"
                    },
                    "animal": {
                        "type": "string"
                    },
                    "last_action_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}